Improve 'Lottery Scheduler' multi-core implementation
Use of the array 'tickets' to make sure the actual winning process of
the lottery gets the CPU in a multi-core environment.
Otherwise, some process could change state from 'RUNNABLE' to 'RUNNING'
in the time between the two 'for' loops and the CPU might end up being
given to the wrong process.